spapr_pci: fix demotion in vfio window size calculation
When we shift tcet->nb_table by tcet->page_shift to calculate
window size, we end up with 0 due to nb_table being uint64_t
and page_shift being uint32_t. This ends up causing the result
to get truncated to uint32_t, which results in 0 for large
windows. This in turn results in VFIO_IOMMU_TCE_CREATE failing
since 0 is an invalid window size.

It's not clear to me why this happens, since C suggests that
the left operand of << defines the result type. Very strange...
but without this we do indeed end up passing a 0 window size
to VFIO_IOMMU_TCE_CREATE, so this is very much needed.

+++ hw/ppc/spapr_pci.c
@@ -731,7 +731,8 @@ static int spapr_phb_dma_update(Object *child, void *opaque)
 
     ret = spapr_phb_vfio_dma_init_window(sphb,
                                          tcet->page_shift,
-                                         tcet->nb_table << tcet->page_shift,
+                                         (uint64_t)tcet->nb_table <<
+                                         tcet->page_shift,
                                          &bus_offset);
     if (ret) {
         return ret;
